@@ -1,10 +1,8 @@
 package com.huehn.initword.activity;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
+
+import androidx.annotation.Nullable;
 
 import com.heaven7.android.dragflowlayout.DragFlowLayout;
 import com.huehn.initword.R;
