@@ -2,8 +2,9 @@ package com.huehn.initword.ui.view.movescale;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.support.annotation.Nullable;
-import android.support.v4.view.GestureDetectorCompat;
+
+import androidx.annotation.Nullable;
+import androidx.core.view.GestureDetectorCompat;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.ViewGroup;
