@@ -1,6 +1,7 @@
 package com.huehn.initword.basecomponent.manager;
 
-import android.support.v4.app.Fragment;
+
+import androidx.fragment.app.Fragment;
 
 public class PermissionFragment extends Fragment {
 
