@@ -1,7 +1,8 @@
 package com.huehn.initword.basecomponent.manager;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
+
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 
 import com.huehn.initword.basecomponent.base.BaseActivity;
 
