@@ -3,7 +3,8 @@ package com.facebook.login;
 import android.content.Context;
 import android.os.Bundle;
 import android.os.Parcel;
-import android.support.v4.app.FragmentActivity;
+
+import androidx.fragment.app.FragmentActivity;
 
 import com.facebook.FacebookException;
 import com.facebook.internal.FacebookDialogFragment;
