@@ -2,8 +2,9 @@
 
 package com.huehn.initword.activity;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.huehn.initword.R;
 
