@@ -1,9 +1,10 @@
 package com.huehn.initword.activity;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
 import android.view.View;
 import android.widget.TextView;
+
+import androidx.annotation.Nullable;
 
 import com.huehn.initword.R;
 import com.huehn.initword.basecomponent.base.BaseActivity;
