@@ -1,16 +1,18 @@
 package com.huehn.initword.ui.view.viewpager;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.view.ViewPager;
+
 import android.util.AttributeSet;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.viewpager.widget.ViewPager;
 
 /**
  * @author huehn
  * 3d滑动首页
  */
-public class GreatViewPager extends ViewPager{
+public class GreatViewPager extends ViewPager {
 
     private IPagerCallback iPagerCallback;
     private GreatViewPagerAdapter greatViewPagerAdapter;
