@@ -1,10 +1,11 @@
 package com.huehn.initword.ui.view.viewpager;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v4.view.PagerAdapter;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.viewpager.widget.PagerAdapter;
 
 import com.huehn.initword.core.utils.Log.LogManager;
 
