@@ -1,10 +1,11 @@
 package com.huehn.initword.ui.view.viewpager;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
 import android.util.AttributeSet;
 import android.view.View;
 import android.widget.LinearLayout;
+
+import androidx.annotation.Nullable;
 
 import com.huehn.initword.R;
 
