@@ -1,12 +1,12 @@
 package com.huehn.initword.view
 
 import android.content.Context
-import android.support.v7.widget.Toolbar
 import android.util.AttributeSet
 import android.widget.ImageView
+import android.widget.Toolbar
 import com.huehn.initword.R
 
-class ScrollingToolBar: Toolbar{
+class ScrollingToolBar: Toolbar {
     private lateinit var mContext: Context
 
     constructor(context: Context): this(context, null)
