@@ -3,12 +3,13 @@ package com.huehn.initword.basecomponent.base;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.os.PersistableBundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.ActivityCompat;
-import android.support.v7.app.AppCompatActivity;
 import android.util.SparseArray;
 
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
 
 import com.huehn.initword.basecomponent.bean.permission.PermissionResult;
 import com.huehn.initword.core.utils.Exception.AppException;
