@@ -1,9 +1,10 @@
 package com.facebook.login;
 
 import android.content.Intent;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
 import android.widget.Toast;
+
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 
 import com.facebook.FacebookActivity;
 import com.facebook.internal.FacebookDialogFragment;
