@@ -4,13 +4,13 @@ import android.app.Activity;
 import android.content.ComponentName;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.LayoutRes;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import com.huehn.initword.core.utils.Log.LogManager;
 
@@ -111,7 +111,6 @@ public class LoginMyFragment extends Fragment {
         return view;
     }
 
-    @LayoutRes
     protected int getLayoutResId() {
         return R.layout.com_facebook_login_fragment;
     }
